+++ CHANGELOG.md
@@ -2,6 +2,8 @@ The changelog for Bonvoy
 
 ## Pending Release
 
+## 0.2.0
+
 * Add `-o json` to output data into a JSON output format
 * Adjust `server restart` to restart the Nomad allocation rather than killing Envoy.
   This allows for a more graceful degradation of traffic to the service.